Testien kaannoskorjaus: Double-vertailut ja ei nulleja Media3-konstruktoreihin

+++ app/src/test/java/fi/aalto/radio/audio/AudioGainTest.kt
@@ -8,18 +8,18 @@ class AudioGainTest {
 
     @Test
     fun `no adjustment leaves the volume untouched`() {
-        assertEquals(1f, decibelsToVolume(0), 0.001f)
+        assertEquals(1.0, decibelsToVolume(0).toDouble(), 0.001)
     }
 
     @Test
     fun `a boost is not done with the player volume`() {
         // Louder is the loudness enhancer's job; the player cannot exceed 1.
-        assertEquals(1f, decibelsToVolume(6), 0.001f)
+        assertEquals(1.0, decibelsToVolume(6).toDouble(), 0.001)
     }
 
     @Test
     fun `minus six decibels is about half the amplitude`() {
-        assertEquals(0.501f, decibelsToVolume(-6), 0.01f)
+        assertEquals(0.501, decibelsToVolume(-6).toDouble(), 0.01)
     }
 
     @Test
